Rename `Quilt` to `Prov` (continued)
