+++ plugins/nf-prov/src/resources/META-INF/extensions.idx
@@ -1,1 +1,1 @@
-nextflow.quilt.QuiltObserverFactory
+nextflow.prov.ProvObserverFactory
